implement demo test on API connection in table_fragment
