+++ RestaurantApp/.gitignore
@@ -6,4 +6,5 @@
 .DS_Store
 /build
 /captures
+/app/build
 .externalNativeBuild
