+++ RestaurantApp/CLICK_ME_TO_RUN_MOCKUP.bat
@@ -1,0 +1,1 @@
+java -jar spring_mockup_demo-1.0-SNAPSHOT.jar
