+++ RestaurantApp/app/build.gradle
@@ -4,11 +4,13 @@ android {
     compileSdkVersion 26
     defaultConfig {
         applicationId "day01.swomfire.restaurantapp"
-        minSdkVersion 19
+        minSdkVersion 23
         targetSdkVersion 26
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+        multiDexEnabled true //important
     }
     buildTypes {
         release {
@@ -25,11 +27,11 @@ dependencies {
     implementation 'com.android.support:support-v4:26.1.0'
 
     // Retrofit
-    compile 'com.squareup.retrofit2:retrofit:2.1.0'
+    implementation 'com.squareup.retrofit2:retrofit:2.1.0'
 
     // JSON Parsing
-    compile 'com.google.code.gson:gson:2.6.1'
-    compile 'com.squareup.retrofit2:converter-gson:2.1.0'
+    implementation 'com.google.code.gson:gson:2.6.1'
+    implementation 'com.squareup.retrofit2:converter-gson:2.1.0'
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.1'
