+++ RestaurantApp/app/src/main/AndroidManifest.xml
@@ -28,4 +28,6 @@
         </activity>
     </application>
 
+    <uses-permission android:name="android.permission.INTERNET"/>
+
 </manifest>
