+++ RestaurantApp/app/src/main/java/Service/TabHostServiceImpl.java
@@ -1,4 +1,4 @@
-package Service;
+package service;
 
 import android.content.Context;
 import android.support.v4.app.FragmentTabHost;
