+++ RestaurantApp/app/src/main/java/utils/ApiUtils.java
@@ -7,8 +7,8 @@ import data.remote.RetrofitClient;
  * Created by elpsychris on 03/12/2018.
  */
 
-public class ApiUtils {
-    public static final String LOCAL_IP = "http://192.168.2.27";
+public class APIUtils {
+    public static final String LOCAL_IP = "http://192.168.0.3";
     public static final String PORT = "8080";
     public static final String BASE_URL = LOCAL_IP + ":" + PORT;
